layout changes

+++ src/components/DeleteBtn/index.tsx
@@ -1,21 +1,29 @@
 "use client";
 import { useRouter } from "next/navigation";
+import Image from "next/image";
 import deleteNote from "~/serverActions/deleteNote";
 
-const DeleteBtn = ({ noteId }: { noteId: number }) => {
+const DeleteBtn = (props: {
+  noteId: number;
+  className?: string | undefined;
+}) => {
+  const { noteId, className = "" } = props;
   const router = useRouter();
   return (
     <button
-      className="border py-1 px-4 rounded-xl mr-4"
+      className={`${className}`}
       type="button"
       onClick={async () => {
-        alert(`are you sure you want to delete ${noteId}?`);
-        const deletedNoteId = await deleteNote(noteId);
-        alert(`Note ${deletedNoteId} Deleted`);
-        router.push("/");
+        if (confirm(`are you sure you want to delete ${noteId}?`)) {
+          const deletedNoteId = await deleteNote(noteId);
+          alert(`Note ${deletedNoteId} Deleted`);
+          router.push("/");
+        }
       }}
+      title="delete note"
+      aria-label="delete note"
     >
-      Delete
+      <Image src="/delete_52.svg" height="36" width="36" />
     </button>
   );
 };
